requirements.txt
# Core dependencies
typing-extensions>=4.0.0
psutil>=5.9.0
spacy>=3.5.0
nltk>=3.8.0
semver>=3.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=2.0.0

# Backend API dependencies
fastapi>=0.104.0
jsonschema>=4.18.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6
websockets>=11.0.0
PyYAML>=6.0.1

# Cognitive transparency dependencies
networkx>=3.1.0
numpy>=1.24.0,<2.0
scipy>=1.11.0
matplotlib>=3.7.0
plotly>=5.17.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.0.0
mypy>=1.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
